Files for adding human classified Single Ventricle
Adds SV-Gold-Simplified.xlsx explanation below.

GOLD STANDARD.xlsx contains a list of rows that should have SV_GROUP set to 1.
To get this group you had to manual filter by a highlighted color on three different rows.
Since this is not the ideal format for a machine to read, created SV-GOLD-SIMPLIFIED.xlsx
This is the two columns of the GOLD STANDARD for all the patients that should have SV_GROUP set to 1.

Adds sv-add-gold.py

sv-add-gold.py reads the SV-Gold-Simplified.xlsx and updates the SV_GROUP column
from txpl_project.xlsx that should have been set to 1 based on human reviewed data in SV-Gold-Simplified.xlsx.

Adds txpl_project_updates.xlsx

This is the output of sv-add-gold. SV_GROUP values that should be 1 based on textual data that
a human reviewed have been changed from 0 to 1. 101 rows were updated.